is-backend/requirements.txt
gunicorn==23.0.0
flask==3.1.0
flask-cors==5.0.1
openai==1.77.0
python-dotenv==1.1.0
